requirements.txt
discord.py>=2.4.0
pillow>=10.0.0
pytesseract>=0.3.10
motor>=3.6.0
